Typo error fixed.
There is a typo error in the images section. The second line of this section, just above the code snippet is written as, "This value is the used as a lookup in the emoji.img_sets property, which defines each set."

The corrected line could be:
"This value is used as a lookup in the emoji.img_sets property, which defines each set." (Here second 'the' is removed in the line).

Thank you.